Allow multiple calls to build() with the same arguments
Ninja requires that each target have a single associated build
statement, so we have a check at configure time that we're not
making this mistake.  This commit loosens the rules somewhat,
we now allow requesting to generate the same target with more
than one build statement as long as all the build statements
for generating that target are identical.