usrmerge: Fix permissions
We perform usrmerge on a package-by-package basis.  We check the package
out, merge `bin`, `lib` and `sbin` into `usr` and setup symlinks and then
check it back in again.  We use `--devino-canonical` so the permissions
and ownership of the files are not modified when checking back in.  The
same doesn't apply to the directories which would be set to be owned by
the UID/GID of the current user.

This commit fixes that resetting the ownership of all directories to
root:root.  This isn't the ideal, but fixing that will require merging
https://github.com/ostreedev/ostree/pull/1678 .