Fix failure when a deb is available from multiple urls
I saw this with a package that was in universe in Ubuntu 18.04, but in
main in Ubuntu 16.04.  For both targets to co-exist in the same
`ninja.build` we need to explicitly allow this.

This is fine as we check by SHA256 anyway - the target includes the SHA256
of the downloaded deb anyway.